backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic-settings
python-dotenv
requests
beautifulsoup4
pytest
httpx
redis>=5.0.0
python-dateutil>=2.9.0
